requirements.txt
fastapi
uvicorn
chromadb
langchain
langchain-community
gitpython
python-dotenv
requests
tiktoken
langchain-text-splitters
groq
sentence-transformers
networkx